+++ src/main/java/com/navigationasistance/interfaces/LocalidadInterface.java
@@ -1,5 +1,6 @@
 package com.navigationasistance.interfaces;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.navigationasistance.modelo.Localidad;
@@ -7,11 +8,11 @@ import com.navigationasistance.modelo.Localidad;
 public interface LocalidadInterface {
     List<Localidad> listar();
 
-    Localidad listarId(Integer id);
+    Localidad listarId(BigDecimal id);
 
     int addLocalidad(Localidad l);
 
     int updLocalidad(Localidad l);
 
-    int delLocalidad(Integer id);
+    int delLocalidad(BigDecimal id);
 }
