+++ src/main/java/com/navigationasistance/mapper/TypeEventRowMapper.java
@@ -10,9 +10,9 @@ public class TypeEventRowMapper implements RowMapper<TypeEvent>{
     @Override
     public TypeEvent mapRow(ResultSet rs, int rowNum) throws SQLException {
         TypeEvent typeEvent = new TypeEvent();
-        typeEvent.setId(rs.getInt("id"));
+        typeEvent.setId(rs.getBigDecimal("id"));
         typeEvent.setType_nombre(rs.getString("type_nombre"));
-        typeEvent.setType_importancia(rs.getInt("type_importancia"));
+        typeEvent.setType_importancia(rs.getBigDecimal("type_importancia"));
 
         return typeEvent;
     }
